brapi_get_call {QBMS}	R Documentation

Internal Function for Core BrAPI GET Calls

Description

Fetches data from an API endpoint, handles pagination by retrieving all pages, and consolidates the results into a single data frame.

Usage

brapi_get_call(call_url, nested = TRUE, caller_func = NA)

Arguments

call_url
Character string specifying the base URL of the API endpoint to request.

nested
Logical value indicating whether to flatten nested lists in the JSON responses. Defaults to TRUE.

caller_func
Character string identifying the name of the function that invoked brapi_get_call().

Details

This function performs the following steps:

Fetches the first page synchronously to determine the total number of pages.

If multiple pages exist, it asynchronously fetches the remaining pages using get_async_pages().

Consolidates the data from all pages into a single data frame.

Updates global state variables with pagination information.

It relies on global variables from qbms_globals to manage state and configuration.

Value

A list containing the consolidated data and associated metadata from the API response.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

[Package QBMS version 2.0.0 Index]