brapi_post_search_allelematrix {QBMS}	R Documentation

Internal Function Used for Core BrAPI POST Calls (Allele Matrix Search)

Description

This function is used internally to execute POST calls for retrieving the allele matrix via BrAPI. It handles the post request, waits for the results asynchronously if required, and processes the results.

Usage

brapi_post_search_allelematrix(call_url, call_body, nested = TRUE)

Arguments

call_url
BrAPI URL for the POST request.

call_body
The request body to send with the POST request.

nested
Logical indicating whether to flatten the nested structure. Default is TRUE.

Value

A list of results obtained from the BrAPI POST call.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

[Package QBMS version 2.0.0 Index]