build_pedigree_table {QBMS}	R Documentation

Building Pedigree Table Recursively

Description

Recursively builds a pedigree table by extracting and tracking parents for each genotype/germplasm in the provided list. The function handles backcross cases and updates the pedigree data frame with parent information for multiple generations.

Usage

build_pedigree_table(
  geno_list = NULL,
  pedigree_list = NULL,
  pedigree_df = NULL
)

Arguments

geno_list
A character vector of genotype/germplasm names.

pedigree_list
A character vector of associated pedigree strings, corresponding to the genotypes in geno_list.

pedigree_df
A data frame of pedigrees from a previous iteration, used to accumulate pedigree data. If NULL, a new data frame is created.

Value

A data frame with three columns: - 'Variety': The identifier for the individual genotype. - 'Female': The identifier for the female parent. - 'Male': The identifier for the male parent. The pedigree is built recursively, with individuals listed before any appearance as a parent.

Author(s)

Khaled Al-Shamaa, k.el-shamaa@cgiar.org

[Package QBMS version 2.0.0 Index]