Calculate the Bioclimatic Variables

This function calculates the 19 standard bioclimatic variables derived from monthly temperature and precipitation data. Bioclimatic variables are often used in ecological modeling and species distribution modeling to capture biologically meaningful patterns in climate data, including annual trends, seasonality, and extreme environmental factors.

The bioclimatic variables represent metrics such as the annual mean temperature, temperature seasonality, and precipitation patterns (e.g., wettest or driest quarter). These metrics help to model species distributions and analyze ecological dynamics.

The bioclimatic variables are coded as follows:

BIO1 = Annual Mean Temperature

BIO2 = Mean Diurnal Range (Mean of monthly (max temp - min temp))

BIO3 = Isothermality (BIO2/BIO7) (* 100)

BIO4 = Temperature Seasonality (standard deviation * 100)

BIO5 = Max Temperature of Warmest Month

BIO6 = Min Temperature of Coldest Month

BIO7 = Temperature Annual Range (BIO5 - BIO6)

BIO8 = Mean Temperature of Wettest Quarter

BIO9 = Mean Temperature of Driest Quarter

BIO10 = Mean Temperature of Warmest Quarter

BIO11 = Mean Temperature of Coldest Quarter

BIO12 = Annual Precipitation

BIO13 = Precipitation of Wettest Month

BIO14 = Precipitation of Driest Month

BIO15 = Precipitation Seasonality (Coefficient of Variation)

BIO16 = Precipitation of Wettest Quarter

BIO17 = Precipitation of Driest Quarter

BIO18 = Precipitation of Warmest Quarter

BIO19 = Precipitation of Coldest Quarter

These variables are computed using temperature and precipitation data in a standard format, and are critical for understanding species habitats and the effects of climate on ecosystems.

This work is derived from the dismo R package.

Usage

calc_biovars(data)

Arguments

data

A data frame containing monthly climate data. The data frame must include:

year: The year for each set of monthly data.

ppt: Monthly precipitation values (in mm).

tmin: Monthly minimum temperature values (in degrees Celsius).

tmax: Monthly maximum temperature values (in degrees Celsius).

The data should contain 12 rows (one for each month from January to December) per year, with the columns sorted in the order of year, ppt, tmin, and tmax.

Value

A data frame with 19 columns representing the bioclimatic variables (BIO1 to BIO19) and an additional column for the year. The output data frame provides one row per year, with each column corresponding to one of the bioclimatic variables described above.

References

Nix, 1986. A biogeographic analysis of Australian elapid snakes. In: R. Longmore (ed.). Atlas of elapid snakes of Australia. Australian Flora and Fauna Series 7. Australian Government Publishing Service, Canberra.

Author

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

Robert Hijmans, Museum of Vertebrate Zoology, UC Berkeley