get_brapi_url {QBMS}	R Documentation

Get the BrAPI Endpoint URL for a given QBMS function

Description

Constructs the BrAPI endpoint URL for a given QBMS function based on the configured server, crop, and BrAPI version. The function name is mapped to the corresponding BrAPI call using internal mapping.

Usage

get_brapi_url(func_name)

Arguments

func_name
(string) The name of the QBMS function for which the BrAPI endpoint URL is required.

Value

A string representing the BrAPI endpoint URL.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

[Package QBMS version 2.0.0 Index]