get_germplasm_attributes {QBMS}	R Documentation

Retrieve Attributes for a Specified Germplasm

Description

Retrieves a detailed list of attributes for a given germplasm, such as its origin, donors, and taxonomic information.

Usage

get_germplasm_attributes(germplasm_name = "")

Arguments

germplasm_name
The name of the germplasm.

Value

A data frame containing the attributes associated with the specified germplasm.

Author(s)

Johan Steven Aparicio, j.aparicio@cgiar.org

See Also

login, set_crop, get_germplasm_data

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()
  set_crop("wheat")
  germplasm_attributes <- get_germplasm_attributes("Jabal")
}

[Package QBMS version 2.0.0 Index]