get_germplasm_data {QBMS}	R Documentation

Retrieve Observations Data for a Specified Germplasm.

Description

Retrieves all available observations data for the given germplasm in the current active crop. This data is aggregated across all trials in the crop database.

Usage

get_germplasm_data(germplasm_name = "")

Arguments

germplasm_name
The name of the germplasm.

Value

A data frame containing all available observations data for the specified germplasm.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, get_germplasm_attributes

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()
  set_crop("wheat")
  germplasm_observations <- get_germplasm_data("Jabal")
  head(germplasm_observations)
}

[Package QBMS version 2.0.0 Index]