get_germplasm_list {QBMS}	R Documentation

Get the Germplasm List of the Current Active Study

Description

Retrieves the list of germplasm (genetic material) used in the currently active study, which must be set using the set_study function.

Usage

get_germplasm_list()

Value

A data frame containing the germplasm list for the active study.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, set_trial, set_study for related operations on crops and studies.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()
  set_crop("wheat")
  set_program("Wheat International Nurseries")
  set_trial("IDYT39")
  set_study("IDYT39 Environment Number 9")
  germplasm <- get_germplasm_list()
  head(germplasm)
}

[Package QBMS version 2.0.0 Index]