get_login_details {QBMS}	R Documentation

Login Pop-Up Window

Description

Opens a GUI pop-up window using Tcl/Tk to prompt the user for their username and password. The window title and prompt message adapt based on the type of server being used (e.g., BMS, GIGWA).

Usage

get_login_details()

Value

A vector containing the inserted username and password, with names usr and pwd respectively.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

[Package QBMS version 2.0.0 Index]