get_parents {QBMS}	R Documentation

Get Direct Parents

Description

Utility function to split a given pedigree string and retrieve the pedigrees of the direct parents (female and male). The function handles different formats of cross representations, such as single slashes (/), double slashes (//), or numbered crosses (e.g., /3/). It extracts the highest cross order when available and returns the sub-pedigree for the immediate parents.

Usage

get_parents(pedigree)

Arguments

pedigree
A string providing the parentage through which a cultivar was obtained.

Value

A vector of two items representing the direct female and male parents. If parent information is unavailable or unknown, 'NA' is returned for the respective parent.

Author(s)

Khaled Al-Shamaa, k.el-shamaa@cgiar.org

[Package QBMS version 2.0.0 Index]