get_program_studies {QBMS}	R Documentation

Get the List of Trials, Studies, and Locations Information for the Current Selected Program

Description

Retrieves comprehensive information about the trials, studies, and environments/locations within the current active breeding program, as configured in the internal state object using the set_program function. This includes test and check entry counts for each study.

Usage

get_program_studies()

Value

A data frame containing detailed information for each study within the program's trials, including trial names, study names, location information, and entry counts.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()
  set_crop("wheat")
  set_program("Wheat International Nurseries")
  program_studies <- get_program_studies()
  head(program_studies)
}

[Package QBMS version 2.0.0 Index]