get_program_trials {QBMS}	R Documentation

Retrieve the List of Trials for the Active Breeding Program

Description

This internal function retrieves the list of trials for the currently active breeding program and crop combination. The crop and program must be set using set_crop and set_program prior to calling this function.

Usage

get_program_trials()

Value

A data frame containing information on trials for the active breeding program.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, list_trials

[Package QBMS version 2.0.0 Index]