get_study_info {QBMS}	R Documentation

Get the Details/Metadata of the Current Active Study

Description

Retrieves detailed metadata for the currently active study. The study must be set using the set_study function, and its details will be fetched from the BrAPI server.

Usage

get_study_info()

Value

A data frame containing the metadata of the active study. Returns NULL if no study metadata is available.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, set_trial, set_study for related crop and study management.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()
  set_crop("wheat")
  set_program("Wheat International Nurseries")
  set_trial("IDYT39")
  set_study("IDYT39 Environment Number 9")
  info <- get_study_info()
}

[Package QBMS version 2.0.0 Index]