get_trial_data {QBMS}	R Documentation

Get the Observations Data of the Current Active Trial

Description

Retrieves the combined observations data (including all studies) for the current active trial, as configured in the internal state object using the set_trial function. This function iterates over all studies within the active trial and aggregates their observation data.

Usage

get_trial_data()

Value

A data frame containing the combined observations data from all studies in the active trial.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, set_trial

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()
  set_crop("wheat")
  set_program("Wheat International Nurseries")
  set_trial("IDYT39")
  MET <- get_trial_data()
  head(MET)
}

[Package QBMS version 2.0.0 Index]