get_variants {QBMS}	R Documentation

Get Marker Matrix from the Selected Variant Set

Description

Get a two-dimensional marker matrix for all samples in the currently active variant set. This function is used as an alternative to the get_variantset() function when the server does not support it.

Note: This approach is significantly slower than get_variantset because it fetches data page by page through the BrAPI API, rather than downloading the full matrix in one Falpjack format file.

Usage

get_variants()

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, set_variantset

[Package QBMS version 2.0.0 Index]