get_variantset {QBMS}	R Documentation

Get Marker Matrix from the Selected Variant Set

Description

Downloads a two-dimensional marker matrix for all samples in the currently active variant set, using the efficient Flapjack file format from the server. Row names correspond to marker IDs and column names correspond to unique sample IDs. Genotype calls are returned in character format.

Usage

get_variantset()

Value

A data frame of genotype calls in character format, with markers as rows and samples as columns.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, set_variantset

[Package QBMS version 2.0.0 Index]