gigwa_get_metadata {QBMS}	R Documentation

Get the Metadata of the Current Active GIGWA Run

Description

Retrieve metadata associated with the samples in the current active run, set using the gigwa_set_run() function. The metadata provides additional information about the samples in the selected run.

Usage

gigwa_get_metadata()

Value

A data frame containing metadata attributes for each sample in the active run.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, gigwa_set_run

Examples

if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_set_db("3kG_10M")
  gigwa_set_project("3003_ind")
  gigwa_set_run("1")
  metadata <- gigwa_get_metadata()
}

[Package QBMS version 2.0.0 Index]