gigwa_list_dbs {QBMS}	R Documentation

List GIGWA Databases

Description

Retrieve the list of available databases from the connected GIGWA server. An active connection is required. If not connected, the function will throw an error.

Usage

gigwa_list_dbs()

Value

A list of databases available on the connected GIGWA server.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config

Examples

if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_list_dbs()
}

[Package QBMS version 2.0.0 Index]