gigwa_list_projects {QBMS}	R Documentation

Get the List of All Projects in the Selected GIGWA Database

Description

Retrieve the list of projects available in the currently active GIGWA database, set using gigwa_set_db(). If no database is selected, the function will throw an error.

Usage

gigwa_list_projects()

Value

A list of project names in the selected database.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, gigwa_set_db

Examples

if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_set_db("Sorghum-JGI_v1")
  gigwa_list_projects()
}

[Package QBMS version 2.0.0 Index]