gigwa_list_runs {QBMS}	R Documentation

Get the List of the Run Names Available in the Selected GIGWA Project

Description

Retrieve the list of available runs in the currently active GIGWA project, set using gigwa_set_project(). If no project is selected, an error will be raised.

Usage

gigwa_list_runs()

Value

A list of run names associated with the selected project.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, gigwa_set_project

Examples

if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_set_db("Sorghum-JGI_v1")
  gigwa_set_project("Nelson_et_al_2011")
  gigwa_list_runs()
}

[Package QBMS version 2.0.0 Index]