gigwa_set_db {QBMS}	R Documentation

Set the Current Active GIGWA Database by Name

Description

Select a GIGWA database as the active database for subsequent operations. This updates the internal configuration object and resets any previously selected projects or runs.

Usage

gigwa_set_db(db_name)

Arguments

db_name
The name of the database to set as active.

Value

No return value. Updates the internal configuration with the selected database.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, gigwa_list_dbs

Examples

if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_set_db("Sorghum-JGI_v1")
}

[Package QBMS version 2.0.0 Index]