gigwa_set_run {QBMS}	R Documentation

Set the Current Active GIGWA Run

Description

Select a run from the active GIGWA project and set it as the current active run in the internal state, enabling further data retrieval operations.

Usage

gigwa_set_run(run_name)

Arguments

run_name
The name of the run to set as active.

Value

No return value. Updates the internal state with the selected run.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, gigwa_set_project, gigwa_list_runs

Examples

if (interactive()) {
  set_qbms_config("https://gigwa.southgreen.fr/gigwa/", 
                  time_out = 300, engine = "gigwa", no_auth = TRUE)
  gigwa_set_db("Sorghum-JGI_v1")
  gigwa_set_project("Nelson_et_al_2011")
  gigwa_set_run("run1")
}

[Package QBMS version 2.0.0 Index]