list_crops {QBMS}	R Documentation

Retrieve Supported Crops from the Server

Description

Retrieves the list of crops supported by the connected server. If the crop list is cached in the internal state, it returns the cached data; otherwise, it sends a BrAPI GET request to fetch the crop list.

Usage

list_crops()

Value

A character vector containing the names of supported crops.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop to configure and set the current active crop.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()  # Log in to the server
  list_crops()  # Retrieve list of supported crops
}

[Package QBMS version 2.0.0 Index]