list_locations {QBMS}	R Documentation

Get the List of Locations Information of the Current Selected Crop

Description

Retrieves a list of locations associated with the current active crop, as configured in the internal state object using the set_crop function.

Usage

list_locations()

Value

A data frame containing information about locations relevant to the current crop.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop for related crop operations.

[Package QBMS version 2.0.0 Index]