list_programs {QBMS}	R Documentation

Retrieve Breeding Programs for the Active Crop

Description

Retrieves the list of breeding programs available for the currently selected crop. The crop must be set using the set_crop function prior to calling this.

Usage

list_programs()

Value

A data frame containing the names of breeding programs available for the active crop.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, list_crops for managing server connection and crop selection.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()  # Log in to the server
  set_crop("wheat")  # Set "wheat" as the active crop
  list_programs()  # Retrieve breeding programs for the active crop
}

[Package QBMS version 2.0.0 Index]