list_studies {QBMS}	R Documentation

Get the List of Studies in the Current Active Trial

Description

Retrieves a list of studies (and associated locations) for the currently active trial, as configured using the set_trial function.

Usage

list_studies()

Value

A data frame containing study names and associated location names. If no studies are available, an error is thrown.

Note

This function must be called after a trial has been set using set_trial.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, set_trial for related operations on crops, programs, and trials.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()  # Log in to the server
  set_crop("wheat")  # Set crop
  set_program("Wheat International Nurseries")  # Set breeding program
  set_trial("IDYT39")  # Set trial
  list_studies()  # List studies
}

[Package QBMS version 2.0.0 Index]