list_variantsets {QBMS}	R Documentation

List Variant Sets in the Selected Study

Description

Retrieves the names of all variant sets available in the currently selected study, as set by set_study(). If no study is selected, the function returns an error.

Usage

list_variantsets()

Value

A list of the names of variant sets associated with the selected study.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, set_study

[Package QBMS version 2.0.0 Index]