login {QBMS}	R Documentation

Login to a BrAPI Server

Description

Authenticates with a BrAPI server using a username and password. If credentials are not provided, interactive prompt will request them from the user. This function serves as a wrapper for all login_* authentication methods and accepts additional parameters via ...

Usage

login(username = NULL, password = NULL, ...)

Arguments

username
The username (optional, default is NULL).

password
The password (optional, default is NULL).

...
Additional arguments passed to login_oauth2.

Value

No return value. On success, the access token is stored internally for future use.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login_oauth2

[Package QBMS version 2.0.0 Index]