login_gigwa {QBMS}	R Documentation

Login to the GIGWA Server

Description

Connect to the GIGWA server. If the username or password parameters are missing, a login window will be triggered to capture these details.

All connection settings (server URL, port, API path, and protocol) are read from the qbms_config() list. The function will request an authentication token from the server and update the qbms_state() list with the token.

Usage

login_gigwa(username = NULL, password = NULL)

Arguments

username
The GIGWA username (optional, default is NULL).

password
The GIGWA password (optional, default is NULL).

Value

No return value. The authentication token will be stored internally.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

Examples

if (interactive()) {
  set_qbms_config("http://localhost:59395/gigwa/index.jsp", time_out = 300, engine = "gigwa")

  # Login using your GIGWA account (interactive mode)
  login_gigwa()
  
  # You can pass GIGWA username and password as parameters (batch mode)
  # login_gigwa("gigwadmin", "nimda")
}

[Package QBMS version 2.0.0 Index]