scan_brapi_endpoints {QBMS}	R Documentation

Scan BrAPI Endpoints

Description

Scans the available BrAPI endpoints on the configured source server and checks their accessibility. This function allows users to verify which BrAPI endpoints are available based on the provided IDs.

Usage

scan_brapi_endpoints(programDbId = 0, trialDbId = 0, studyDbId = 0)

Arguments

programDbId
(numeric) The programDbId to scan specific program-related endpoints (default is 0).

trialDbId
(numeric) The trialDbId to scan specific trial-related endpoints (default is 0).

studyDbId
(numeric) The studyDbId to scan specific study-related endpoints (default is 0).

Value

A data frame listing the QBMS function, BrAPI endpoint URL, and availability status for each endpoint.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

[Package QBMS version 2.0.0 Index]