set_crop {QBMS}	R Documentation

Set the Current Active Crop

Description

Updates the internal configuration to set the selected crop as the active one. This must be called before performing crop-specific operations such as retrieving breeding programs.

Usage

set_crop(crop_name)

Arguments

crop_name
A string specifying the name of the crop to set as active.

Value

No return value. The function updates the global state with the selected crop.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, list_crops to validate and retrieve the list of supported crops.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()  # Log in to the server
  set_crop("wheat")  # Set "wheat" as the active crop
}

[Package QBMS version 2.0.0 Index]