set_study {QBMS}	R Documentation

Set the Current Active Study

Description

Updates the internal state to set the selected study as the current active study using the associated studyDbId. This allows operations to be performed within the context of the selected study.

Usage

set_study(study_name)

Arguments

study_name
A string specifying the name of the study to set as active.

Value

No return value. The internal state is updated with the selected study.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, set_trial, list_studies for related operations on crops, programs, trials, and studies.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()  # Log in to the server
  set_crop("wheat")  # Set crop
  set_program("Wheat International Nurseries")  # Set breeding program
  set_trial("IDYT39")  # Set trial
  set_study("IDYT39 Environment Number 9")  # Set study
}
 

[Package QBMS version 2.0.0 Index]