set_trial {QBMS}	R Documentation

Set the Current Active Trial

Description

Updates the internal state to set the selected trial as the current active trial using the associated trialDbId. This enables operations to be carried out within the context of the selected trial.

Usage

set_trial(trial_name)

Arguments

trial_name
A string specifying the name of the trial to set as active.

Value

No return value. The internal state is updated with the selected trial.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

login, set_crop, set_program, list_trials for operations involving crops, programs, and trials.

Examples

if (interactive()) {
  set_qbms_config("https://bms.icarda.org/ibpworkbench")
  login_bms()  # Log in to the server
  set_crop("wheat")  # Set crop
  set_program("Wheat International Nurseries")  # Set breeding program
  set_trial("IDYT39")  # Set trial
}

[Package QBMS version 2.0.0 Index]