set_variantset {QBMS}	R Documentation

Set the Active Variant Set

Description

Selects a variant set by name from the currently active study and updates the internal state to make it the active variant set. This selection is required for subsequent data retrieval operations related to variant sets.

Usage

set_variantset(variantset_name)

Arguments

variantset_name
The name of the variant set to be active.

Value

No return value. Updates the internal state with the selected variant set.

Author(s)

Khaled Al-Shamaa (k.el-shamaa@cgiar.org)

See Also

set_qbms_config, set_study, list_variantsets

[Package QBMS version 2.0.0 Index]